+++ provisioner/darwin/darwin_test.go
@@ -1,4 +1,4 @@
-package provisioners
+package darwin
 
 import (
 	"context"
